feat: add htppOnly cookie method for jwt token
