+++ backend/src/config/constants.ts
@@ -1,0 +1,3 @@
+export const COOKIES = {
+    AUTHORIZATION: 'Authorization',
+};
