+++ backend/src/index.ts
@@ -2,7 +2,7 @@ import { Hono } from 'hono'
 import { prismaMiddleware } from './middlewares/prisma'
 import { authMiddleware } from './middlewares/auth'
 import { HonoBindings } from './config/types'
-import { userRouter } from './routes/user'
+import { authRouter } from './routes/auth'
 import { blogRouter } from './routes/blog'
 import { cors } from 'hono/cors'
 
@@ -12,7 +12,7 @@ const app = new Hono<HonoBindings>().basePath('/api/v1')
 app.use('*', prismaMiddleware )
 app.use('*', cors())
 
-app.route('/auth', userRouter)
+app.route('/auth', authRouter)
 app.route('/blog', blogRouter)
